content_sao: Splitted into multiple files: UnitSAO, TestSAO, LuaEntitySAO, LagPool, PlayerSAO and PlayerHPChangeReason.
