+++ src/database/database-files.cpp
@@ -20,7 +20,9 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include <cassert>
 #include <json/json.h>
 #include "database-files.h"
-#include "content_sao.h"
+#include "server/object/LuaEntitySAO.hpp"
+#include "server/object/LagPool.hpp"
+#include "server/object/PlayerSAO.hpp"
 #include "remoteplayer.h"
 #include "settings.h"
 #include "porting.h"
