+++ src/engine/server/object/LuaEntitySAO.hpp
@@ -1,0 +1,93 @@
+/*
+Minetest
+Copyright (C) 2010-2013 celeron55, Perttu Ahola <celeron55@gmail.com>
+
+This program is free software; you can redistribute it and/or modify
+it under the terms of the GNU Lesser General Public License as published by
+the Free Software Foundation; either version 2.1 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Lesser General Public License for more details.
+
+You should have received a copy of the GNU Lesser General Public License along
+with this program; if not, write to the Free Software Foundation, Inc.,
+51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+*/
+
+#pragma once
+
+#include "server/object/UnitSAO.hpp"
+
+/*
+	LuaEntitySAO needs some internals exposed.
+*/
+
+class LuaEntitySAO : public UnitSAO
+{
+public:
+	LuaEntitySAO(ServerEnvironment *env, v3f pos,
+	             const std::string &name, const std::string &state);
+	~LuaEntitySAO();
+	ActiveObjectType getType() const
+	{ return ACTIVEOBJECT_TYPE_LUAENTITY; }
+	ActiveObjectType getSendType() const
+	{ return ACTIVEOBJECT_TYPE_GENERIC; }
+	virtual void addedToEnvironment(u32 dtime_s);
+	static ServerActiveObject* create(ServerEnvironment *env, v3f pos,
+			const std::string &data);
+	void step(float dtime, bool send_recommended);
+	std::string getClientInitializationData(u16 protocol_version);
+	bool isStaticAllowed() const
+	{ return m_prop.static_save; }
+	void getStaticData(std::string *result) const;
+	int punch(v3f dir,
+			const ToolCapabilities *toolcap=NULL,
+			ServerActiveObject *puncher=NULL,
+			float time_from_last_punch=1000000);
+	void rightClick(ServerActiveObject *clicker);
+	void setPos(const v3f &pos);
+	void moveTo(v3f pos, bool continuous);
+	float getMinimumSavedMovement();
+	std::string getDescription();
+	void setHP(s16 hp, const PlayerHPChangeReason &reason);
+	s16 getHP() const;
+	/* LuaEntitySAO-specific */
+	void setVelocity(v3f velocity);
+	void addVelocity(v3f velocity)
+	{
+		m_velocity += velocity;
+	}
+	v3f getVelocity();
+	void setAcceleration(v3f acceleration);
+	v3f getAcceleration();
+
+	void setTextureMod(const std::string &mod);
+	std::string getTextureMod() const;
+	void setSprite(v2s16 p, int num_frames, float framelength,
+			bool select_horiz_by_yawpitch);
+	std::string getName();
+	bool getCollisionBox(aabb3f *toset) const;
+	bool getSelectionBox(aabb3f *toset) const;
+	bool collideWithObjects() const;
+private:
+	std::string getPropertyPacket();
+	void sendPosition(bool do_interpolate, bool is_movement_end);
+
+	std::string m_init_name;
+	std::string m_init_state;
+	bool m_registered = false;
+
+	v3f m_velocity;
+	v3f m_acceleration;
+
+	float m_last_sent_yaw = 0.0f;
+	v3f m_last_sent_position;
+	v3f m_last_sent_velocity;
+	float m_last_sent_position_timer = 0.0f;
+	float m_last_sent_move_precision = 0.0f;
+	std::string m_current_texture_modifier = "";
+};
+
