+++ src/engine/server/object/TestSAO.hpp
@@ -1,0 +1,54 @@
+/*
+Minetest
+Copyright (C) 2010-2013 celeron55, Perttu Ahola <celeron55@gmail.com>
+
+This program is free software; you can redistribute it and/or modify
+it under the terms of the GNU Lesser General Public License as published by
+the Free Software Foundation; either version 2.1 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Lesser General Public License for more details.
+
+You should have received a copy of the GNU Lesser General Public License along
+with this program; if not, write to the Free Software Foundation, Inc.,
+51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+*/
+
+#pragma once
+
+#include "serverobject.h"
+#include "constants.h"
+
+class TestSAO : public ServerActiveObject
+{
+public:
+	TestSAO(ServerEnvironment *env, v3f pos):
+		ServerActiveObject(env, pos)
+	{
+		ServerActiveObject::registerType(getType(), create);
+	}
+	ActiveObjectType getType() const
+	{ return ACTIVEOBJECT_TYPE_TEST; }
+
+	static ServerActiveObject* create(ServerEnvironment *env, v3f pos,
+			const std::string &data)
+	{
+		return new TestSAO(env, pos);
+	}
+
+	void step(float dtime, bool send_recommended);
+
+	bool getCollisionBox(aabb3f *toset) const { return false; }
+
+	virtual bool getSelectionBox(aabb3f *toset) const { return false; }
+
+	bool collideWithObjects() const { return false; }
+
+private:
+	float m_timer1 = 0;
+	float m_age = 0;
+};
+
